feat(signUp-material) deleted original auth-form.js file, replaced with new AuthForm.js, integrating material-ui
