+++ client/components/index.js
@@ -5,5 +5,5 @@
  */
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
+export {Login, Signup} from './AuthForm'
 export {Checkout} from './checkout-form'
