+++ client/routes.js
@@ -2,12 +2,12 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {Login, Signup} from './components'
 import {me} from './store'
 import SingleProduct from './components/SingleProduct'
 import AllProducts from './components/AllProducts'
 import Checkout from './components/checkout-form'
 import Cart from './components/cart'
+import {SignUp, LogIn} from './components/AuthForm'
 
 /**
  * COMPONENT
@@ -24,8 +24,8 @@ class Routes extends Component {
         {/* Routes placed here are available to all visitors */}
         <Route exact path="/products/:id" component={SingleProduct} />
         <Route exact path="/products" component={AllProducts} />
-        <Route path="/login" component={Login} />
-        <Route path="/signup" component={Signup} />
+        <Route path="/login" component={LogIn} />
+        <Route path="/signup" component={SignUp} />
         <Route path="/checkout" component={Checkout} />
         <Route path="/cart" component={Cart} />
         {isLoggedIn && (
